+++ angular-workspace/projects/ni/nimble-angular/src/directives/text-area/tests/nimble-text-area.directive.spec.ts
@@ -1,8 +1,7 @@
 import { Component, ElementRef, ViewChild } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { TextAreaAppearance, TextAreaResize } from '@ni/nimble-components/dist/esm/text-area/types';
 import type { BooleanValueOrAttribute, NumberValueOrAttribute } from '../../utilities/template-value-helpers';
-import { NimbleTextAreaDirective, TextArea } from '../nimble-text-area.directive';
+import { NimbleTextAreaDirective, TextArea, TextAreaAppearance, TextAreaResize } from '../nimble-text-area.directive';
 import { NimbleTextAreaModule } from '../nimble-text-area.module';
 
 describe('Nimble text area', () => {
@@ -45,8 +44,8 @@ describe('Nimble text area', () => {
         });
 
         it('has expected defaults for appearance', () => {
-            expect(directive.appearance).toBe(TextAreaAppearance.Outline);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Outline);
+            expect(directive.appearance).toBe(TextAreaAppearance.outline);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.outline);
         });
 
         it('has expected defaults for readOnly', () => {
@@ -114,7 +113,7 @@ describe('Nimble text area', () => {
         @Component({
             template: `
                 <nimble-text-area #textArea
-                    appearance="${TextAreaAppearance.Block}"
+                    appearance="${TextAreaAppearance.block}"
                     readonly
                     resize="${TextAreaResize.horizontal}"
                     autofocus
@@ -150,8 +149,8 @@ describe('Nimble text area', () => {
         });
 
         it('will use template string values for appearance', () => {
-            expect(directive.appearance).toBe(TextAreaAppearance.Block);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Block);
+            expect(directive.appearance).toBe(TextAreaAppearance.block);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.block);
         });
 
         it('will use template string values for readOnly', () => {
@@ -237,7 +236,7 @@ describe('Nimble text area', () => {
         class TestHostComponent {
             @ViewChild('textArea', { read: NimbleTextAreaDirective }) public directive: NimbleTextAreaDirective;
             @ViewChild('textArea', { read: ElementRef }) public elementRef: ElementRef<TextArea>;
-            public appearance = TextAreaAppearance.Block;
+            public appearance: TextAreaAppearance = TextAreaAppearance.block;
             public readonly = false;
             public resize: TextAreaResize = TextAreaResize.horizontal;
             public autofocus = false;
@@ -268,14 +267,14 @@ describe('Nimble text area', () => {
         });
 
         it('can be configured with property binding for appearance', () => {
-            expect(directive.appearance).toBe(TextAreaAppearance.Block);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Block);
+            expect(directive.appearance).toBe(TextAreaAppearance.block);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.block);
 
-            fixture.componentInstance.appearance = TextAreaAppearance.Outline;
+            fixture.componentInstance.appearance = TextAreaAppearance.outline;
             fixture.detectChanges();
 
-            expect(directive.appearance).toBe(TextAreaAppearance.Outline);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Outline);
+            expect(directive.appearance).toBe(TextAreaAppearance.outline);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.outline);
         });
 
         it('can be configured with property binding for readOnly', () => {
@@ -433,7 +432,7 @@ describe('Nimble text area', () => {
         class TestHostComponent {
             @ViewChild('textArea', { read: NimbleTextAreaDirective }) public directive: NimbleTextAreaDirective;
             @ViewChild('textArea', { read: ElementRef }) public elementRef: ElementRef<TextArea>;
-            public appearance: TextAreaAppearance = TextAreaAppearance.Block;
+            public appearance: TextAreaAppearance = TextAreaAppearance.block;
             public readonly: BooleanValueOrAttribute = null;
             public resize: TextAreaResize = TextAreaResize.horizontal;
             public autofocus: BooleanValueOrAttribute = null;
@@ -464,14 +463,14 @@ describe('Nimble text area', () => {
         });
 
         it('can be configured with attribute binding for appearance', () => {
-            expect(directive.appearance).toBe(TextAreaAppearance.Block);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Block);
+            expect(directive.appearance).toBe(TextAreaAppearance.block);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.block);
 
-            fixture.componentInstance.appearance = TextAreaAppearance.Outline;
+            fixture.componentInstance.appearance = TextAreaAppearance.outline;
             fixture.detectChanges();
 
-            expect(directive.appearance).toBe(TextAreaAppearance.Outline);
-            expect(nativeElement.appearance).toBe(TextAreaAppearance.Outline);
+            expect(directive.appearance).toBe(TextAreaAppearance.outline);
+            expect(nativeElement.appearance).toBe(TextAreaAppearance.outline);
         });
 
         it('can be configured with attribute binding for readOnly', () => {