+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleMenuItem.razor.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.AspNetCore.Components;
+
+namespace NimbleBlazor;
+
+public partial class NimbleMenuItem : ComponentBase
+{
+    [Parameter]
+    public bool? Disabled { get; set; }
+
+    [Parameter]
+    public bool? Checked { get; set; }
+
+    [Parameter]
+    public RenderFragment? ChildContent { get; set; }
+
+    [Parameter(CaptureUnmatchedValues = true)]
+    public IDictionary<string, object>? AdditionalAttributes { get; set; }
+}