+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleToolbar.razor.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Components;
+
+namespace NimbleBlazor;
+
+public partial class NimbleToolbar
+{
+    [Parameter]
+    public RenderFragment? ChildContent { get; set; }
+}