+++ packages/nimble-blazor/NimbleBlazor/Components/NimbleTreeView.razor.cs
@@ -1,0 +1,18 @@
+﻿using Microsoft.AspNetCore.Components;
+
+namespace NimbleBlazor;
+
+public partial class NimbleTreeView : ComponentBase
+{
+    [Parameter]
+    public RenderFragment? ChildContent { get; set; }
+
+    [Parameter]
+    public bool? RenderCollapsedNodes { get; set; }
+
+    [Parameter]
+    public SelectionMode? SelectionMode { get; set; }
+
+    [Parameter(CaptureUnmatchedValues = true)]
+    public IDictionary<string, object>? AdditionalAttributes { get; set; }
+}