+++ packages/nimble-components/src/patterns/button/styles.ts
@@ -128,12 +128,16 @@ export const styles = css`
     }
 
     :host([disabled]) slot[name='start']::slotted(*) {
-        opacity: 0.6;
+        opacity: 0.3;
     }
 
     [part='end'] {
         display: contents;
         ${iconColor.cssCustomProperty}: ${buttonLabelFontColor};
+    }
+
+    :host([disabled]) slot[name='end']::slotted(*) {
+        opacity: 0.3;
     }
 
     :host([content-hidden]) [part='end'] {
@@ -143,7 +147,7 @@ export const styles = css`
     // prettier-ignore
     .withBehaviors(
         appearanceBehavior(
-            ButtonAppearance.Outline,
+            ButtonAppearance.outline,
             css`
                 .control {
                     background-color: transparent;
@@ -172,7 +176,7 @@ export const styles = css`
             `
         ),
         appearanceBehavior(
-            ButtonAppearance.Ghost,
+            ButtonAppearance.ghost,
             css`
                 .control {
                     background-color: transparent;
@@ -201,7 +205,7 @@ export const styles = css`
             `
         ),
         appearanceBehavior(
-            ButtonAppearance.Block,
+            ButtonAppearance.block,
             css`
                 .control {
                     background-color: rgba(${borderRgbPartialColor}, 0.1);