+++ packages/nimble-components/src/patterns/button/tests/types.spec.ts
@@ -1,0 +1,9 @@
+import type { ButtonAppearance } from '../types';
+
+describe('Button type', () => {
+    it('ButtonAppearance fails compile if assigning arbitrary string values', () => {
+        // @ts-expect-error This expect will fail if the enum-like type is missing "as const"
+        const appearance: ButtonAppearance = 'hello';
+        expect(appearance).toEqual('hello');
+    });
+});